Move included plugins to plugins.included
