+++ xpublish/plugins/hooks.py
