+++ xpublish/plugins/manage.py
